+++ crates/uzumaki/js/constants.ts
@@ -1,1 +1,106 @@
 export const __DEV__ = true; // todo switch with a real one idk
+
+export const INPUT_ATTR_NAMES = new Set([
+  'value',
+  'placeholder',
+  'disabled',
+  'maxLength',
+  'multiline',
+  'secure',
+]);
+export const CHECKBOX_ATTR_NAMES = new Set(['checked']);
+
+export const STYLE_ATTRIBUTE_NAMES = new Set([
+  'w',
+  'h',
+  'minW',
+  'minH',
+  'p',
+  'px',
+  'py',
+  'pt',
+  'pb',
+  'pl',
+  'pr',
+  'm',
+  'mx',
+  'my',
+  'mt',
+  'mb',
+  'ml',
+  'mr',
+  'flex',
+  'flexDir',
+  'flexGrow',
+  'flexShrink',
+  'items',
+  'justify',
+  'gap',
+  'bg',
+  'color',
+  'fontSize',
+  'fontWeight',
+  'rounded',
+  'roundedTL',
+  'roundedTR',
+  'roundedBR',
+  'roundedBL',
+  'border',
+  'borderTop',
+  'borderRight',
+  'borderBottom',
+  'borderLeft',
+  'borderColor',
+  'opacity',
+  'display',
+  'cursor',
+  'hover:bg',
+  'hover:color',
+  'hover:opacity',
+  'hover:borderColor',
+  'active:bg',
+  'active:color',
+  'active:opacity',
+  'active:borderColor',
+  'scrollable',
+  'selectable',
+  'visibility',
+  'overflowWrap',
+  'wordBreak',
+  'position',
+  'top',
+  'right',
+  'bottom',
+  'left',
+  'translate',
+  'translateX',
+  'translateY',
+  'rotate',
+  'scale',
+  'scaleX',
+  'scaleY',
+  'hover:translateX',
+  'hover:translate',
+  'hover:translateY',
+  'hover:rotate',
+  'hover:scale',
+  'hover:scaleX',
+  'hover:scaleY',
+  'active:translateX',
+  'active:translate',
+  'active:translateY',
+  'active:rotate',
+  'active:scale',
+  'active:scaleX',
+  'active:scaleY',
+]);
+
+export const INTRINSIC_ELEMENTS = new Set([
+  'view',
+  'text',
+  'input',
+  'checkbox',
+  'image',
+  'button',
+  /* 'canvas' */ // todo
+]);
