+++ crates/uzumaki/js/core.ts
@@ -11,7 +11,7 @@ interface Core {
   getRootNodeId(windowId: number): NodeId;
   createElement(windowId: number, elementType: string): NodeId;
   createTextNode(windowId: number, text: string): NodeId;
-  setImageData(windowId: number, nodeId: NodeId, data: Uint8Array): void;
+  setEncodedImageData(windowId: number, nodeId: NodeId, data: Uint8Array): void;
   clearImageData(windowId: number, nodeId: NodeId): void;
   appendChild(windowId: number, parentId: NodeId, childId: NodeId): void;
   insertBefore(
@@ -79,3 +79,26 @@ const core: Core = (globalThis as unknown as any)
   .__uzumaki_ops_dont_touch_this__;
 
 export default core;
+
+export function setNativeProp(
+  windowId: number,
+  nodeId: any,
+  propName: string,
+  value: any,
+): void {
+  if (typeof value === 'boolean') {
+    core.setBoolAttribute(windowId, nodeId, propName, value);
+  } else if (typeof value === 'number') {
+    core.setNumberAttribute(windowId, nodeId, propName, value);
+  } else {
+    core.setStrAttribute(windowId, nodeId, propName, String(value));
+  }
+}
+
+export function clearNativeProp(
+  windowId: number,
+  nodeId: any,
+  propName: string,
+): void {
+  core.clearAttribute(windowId, nodeId, propName);
+}
