+++ crates/uzumaki/js/types.ts
@@ -1,1 +1,7 @@
 export type NodeId = number;
+
+export interface ListenerEntry {
+  name: string;
+  handler: Function;
+  capture: boolean;
+}
