+++ crates/uzumaki/package.json
@@ -5,6 +5,7 @@
   "description": "React Desktop UI Framework — native GPU-rendered apps with JSX",
   "license": "MIT OR Apache-2.0",
   "devDependencies": {
+    "@types/deno": "^2.5.0",
     "@types/react": "^19.2.14",
     "@types/react-reconciler": "^0.33.0"
   },
@@ -31,8 +32,7 @@
     "vello"
   ],
   "dependencies": {
-    "react-reconciler": "^0.33.0",
-    "@types/node": "^25.6.0"
+    "react-reconciler": "^0.33.0"
   },
   "repository": {
     "type": "git",
